Refactoring, optimization and new features: Addition of multiple levels of map generation, to create more realistic maps 	modified:   Interpolation.cs 	new file:   MapGenerator.cs 	new file:   Point.cs 	modified:   Program.cs 	modified:   bin/Debug/net7.0/Perlin Map.dll 	modified:   bin/Debug/net7.0/Perlin Map.pdb 	modified:   obj/Debug/net7.0/Perlin Map.GeneratedMSBuildEditorConfig.editorconfig 	modified:   obj/Debug/net7.0/Perlin Map.assets.cache 	modified:   obj/Debug/net7.0/Perlin Map.dll 	modified:   obj/Debug/net7.0/Perlin Map.pdb 	modified:   obj/Debug/net7.0/ref/Perlin Map.dll 	modified:   obj/Debug/net7.0/refint/Perlin Map.dll

+++ Interpolation.cs
@@ -2,7 +2,7 @@ namespace SimplexNoise
 {
     static class Interpolation
     {
-        public static double sumOctave(int num_iterations, int x, int y, float persistence, float scale, int low, int high)
+        public static double sumOctave(int num_iterations, long x, long y, float persistence, float scale, int low, int high)
         {
             var freq = scale;
             //float freq = 1.0f;
